flagged presence of Warped files

+++ UPENN-GBM/normalize_MNI.sh
@@ -39,6 +39,12 @@ for subject in "$BASE_DIR"/UPENN-GBM_data-raw_V3-organized/*; do
             if [[ -f "${OUTPUT_NAME}__1Warp.nii.gz" ]]; then
                 echo "  Transformation already available"
             else
+                ### Perform bias field correction
+                echo "Running N4BiasFieldCorrection on ${SUBJECT_DIR}"
+                N4BiasFieldCorrection -d 3 \
+                    -i "$REFERENCE_IMAGE" \
+                    -o "${OUTPUT_NAME}_biasfieldcorrected.nii.gz"
+
                 MASK_NAME=$(find "$subject" -name "*_segm.nii.gz")
                 if [[ ! -f "$MASK_NAME" ]]; then
                     MASK_NAME=$(find "$subject" -name "*_automated_approx_segm.nii.gz")
@@ -47,23 +53,26 @@ for subject in "$BASE_DIR"/UPENN-GBM_data-raw_V3-organized/*; do
                 fslmaths "$MASK_NAME" -bin "$INVERSE_MASK"
                 ImageMath 3 ${INVERSE_MASK} Neg ${INVERSE_MASK}
 
-                ### Perform bias field correction
-                echo "Running N4BiasFieldCorrection on ${SUBJECT_DIR}"
-                N4BiasFieldCorrection -d 3 \
-                    -i "$REFERENCE_IMAGE" \
-                    -o "${OUTPUT_NAME}_biasfieldcorrected.nii.gz"
-
                 ### Run the antsRegistrationSyN.sh command ###
-                echo "Running antsRegistrationSyNQuick on ${SUBJECT_DIR}"
-                antsRegistrationSyNQuick.sh -d 3 \
-                    -f "${MNI_TEMPLATE}.nii.gz" \
-                    -m "${OUTPUT_NAME}_biasfieldcorrected.nii.gz" \
-                    -o "${OUTPUT_NAME}"__ \
-                    -j 1 \
-                    -x "${MNI_TEMPLATE}_mask.nii.gz" "$INVERSE_MASK" \
-                    -n 4 # Number of threads. BE CAREFUL TO MATCH THE NUMBER OF PROCESSES AND THE NUMBER OF CPU CORES IN THE MACHINE
+                if [[ -f "$MASK_NAME" ]]; then
+                    echo "Running antsRegistrationSyNQuick on ${SUBJECT_DIR} with an inverse lesion mask"
+                    antsRegistrationSyNQuick.sh -d 3 \
+                        -f "${MNI_TEMPLATE}.nii.gz" \
+                        -m "${OUTPUT_NAME}_biasfieldcorrected.nii.gz" \
+                        -o "${OUTPUT_NAME}"__ \
+                        -j 1 \
+                        -x "${MNI_TEMPLATE}_mask.nii.gz" "$INVERSE_MASK" 
 
-                rm "$INVERSE_MASK"
+                    rm "$INVERSE_MASK"
+                else
+                    echo "Running antsRegistrationSyNQuick on ${SUBJECT_DIR} WITHOUT an inverse lesion mask"
+                    antsRegistrationSyNQuick.sh -d 3 \
+                        -f "${MNI_TEMPLATE}.nii.gz" \
+                        -m "${OUTPUT_NAME}_biasfieldcorrected.nii.gz" \
+                        -o "${OUTPUT_NAME}"__ \
+                        -j 1 \
+                        -x "${MNI_TEMPLATE}_mask.nii.gz"
+                fi
                 rm "${OUTPUT_NAME}_biasfieldcorrected.nii.gz"
             fi
